Improve highlights for different line formats
Quickfix lines can come in different formats, depending on what data is
available. Some examples are:

    E file.txt:1:2 this is text
    E file.txt     this is text
      file.txt:1:2 this is text
                   this is text

In all cases we want to ensure that if a file path is present, it's
highlighted. And if a path is not present, we don't want to accidentally
highlight anything like a path. In addition, when only displaying text
we don't want to mess with any whitespace, nor do we want extra leading
whitespace (for alignment).

To achieve this, we start each line with a placeholder _unless_ the item
for that line has a `type` value (e.g. "E" in the above example). These
placeholders are concealed appropriately, and used in the syntax rules
to highlight file paths.

The resulting code isn't great, but it's the best I could come up with.
Functions such as nvim_buf_add_highlight() and nvim_buf_set_extmark()
are not an option, as they require lines/columns to mark/highlight to
already exist. When the `quickfixtextfunc` callback is called, the
quickfix buffer isn't filled in yet, meaning these functions can't be
used. One can sort of work around that by using `vim.schedule()` and set
up highlights later, but this may stop working the moment the ordering
of things change. Based on my experiments it would also require quite a
bit more (ugly) code.

I also experimented with using tags for the various parts of a quickfix
item, and highlight/conceal them. For example, a file path would look
like this:

    <path>foo/bar/baz.txt:1:2</path>

This approach ended up requiring more code, and would require that we
escape these tags in the text displayed for every item. This is prone to
error, so I decided to avoid this.

A downside of the new approach is that if you yank a line from the
quickfix buffer, it includes the concealed characters. In practise I
think this should be fine, as users can always remove the line prefix if
needed.

This fixes https://gitlab.com/yorickpeterse/nvim-pqf/-/issues/3.