Refactor formatting and highlighting
Instead of using a fragile syntax file, we now use vim.highlight.range()
to highlight file paths and line/column numbers properly. In addition,
the code is cleaned up and simplified in various places. Finally,
columns are only displayed if a line is also present, and no
placeholders are displayed if they are missing.

This fixes https://github.com/yorickpeterse/nvim-pqf/issues/9.