Fix highlights when appending to the quickfix
When appending to an existing quickfix window, the start index is
incremented to start at the first new entry. The previous setup would
then incorrectly clear all highlights, and only apply them to the first
existing entry.

This commit changes things around so we only clear highlights when an
entirely new quickfix entry is produced, and to correctly apply
highlights to newly appended items.