Refactor sign text and highlights
The use of :sign-define and related APIs for NeoVim's LSP signs has been
deprecated for a while, and NeoVim hasn't used actual signs for this for
a while now.

Unfortunately, vim.diagnostic doesn't provide the means to _get_
configuration details such as sign highlights and texts, so we still
need to configure those separately.

This commit changes things such that we don't depend on deprecated APIs,
while still allowing users to configure both the sign text and highlight
groups if needed.